refactor: 주소 관련 vuex 로직 분리
